[FLINK-19672][connector-kafka] Update Kafka table sources and sinks
